requirements.txt
discord.py==2.4.0
colorama==0.4.6
PyNaCl==1.5.0
async-timeout==5.0.1
yt_dlp==2024.11.18
cogwatch==3.3.1
httpx==0.27.2
requests==2.32.3
mysql-connector-python==8.3.0